requirements.txt
# requirements.txt
flask==2.3.3
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-semantic-conventions==0.39b0
requests==2.31.0